Added files created thus far.
Since a lot of features are going to be included in this project, it is best to start with version controlling to ensure a copy of codebase.
As of this commit, the code is not functioning and only has lexical analyzer partially implemented.
Going forward will be complete implementation of the analyzer and other processes of compiler desgin.